moved everything in pages to app
