+++ client/src/app/dashboard/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { signOut} from 'next-auth/react';
 import React from 'react'
